Moved CommunicatorMessageSent listener to async Singleton
